% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{getgraph}
\alias{getgraph}
\title{Developing descriptive graphs on youtube videos}
\usage{
getgraph(vid)
}
\arguments{
\item{vid}{a character scalar that contains the vedio id}
}
\value{
a json value
}
\description{
for the given video id a graph on likes, dislikes, comments, shares and views on
youtube and facebook
}

